수정: 게시글 CRUD에 auth 적용

+++ src/services/articlesService.js
@@ -1,23 +1,19 @@
 import { prisma } from '../utils/prisma.js';
 
-async function createArticleInDb(title, content) {
+async function createArticleInDb(title, content, userId) {
   return prisma.article.create({
     data: {
       title: title,
       content: content,
+      userId,
     },
   });
 }
 
 async function findArticles({ sort, search, offset, limit }) {
-  const orderBy = {};
-
-  if (sort === 'recent') {
-    orderBy.createdAt = 'desc';
-  }
+  const orderBy = sort === 'resent' ? { createdAt: 'desc' } : { createdAt: 'asc' };
 
   const where = {};
-
   if (search) {
     where.OR = [{ title: { contains: search } }, { content: { contains: search } }];
   }
@@ -51,18 +47,41 @@ async function findArticleById(id) {
       title: true,
       content: true,
       createdAt: true,
+      user: {
+        select: {
+          id: true,
+          nickname: true,
+          email: true,
+        },
+      },
     },
   });
 }
 
-async function updateArticleInDb(id, updateData) {
+async function updateArticleInDb(id, updateData, userId) {
+  const article = await prisma.article.findUniqueOrThrow({ where: { id } });
+
+  if (article.userId !== userId) {
+    const error = new Error('수정 권한이 없습니다.');
+    error.status = 403;
+    throw error;
+  }
+
   return prisma.article.update({
     where: { id },
     data: updateData,
   });
 }
 
-async function deleteArticleInDb(id) {
+async function deleteArticleInDb(id, userId) {
+  const article = await prisma.article.findUniqueOrThrow({ where: { id } });
+
+  if (article.userId !== userId) {
+    const error = new Error('삭제 권한이 없습니다.');
+    error.status = 403;
+    throw error;
+  }
+
   return prisma.article.delete({
     where: { id },
   });
